rearrange to lib and log
